Struct soap_rs::wsdl::WsdlDocument

pub struct WsdlDocument {
    pub parsed: bool,
    pub version: WsdlVersion,
    pub types: HashMap<String, WsdlType>,
    /* private fields */
}

A structure to describe what services are available according to the provided WSDL document

Fields
parsed: bool

Stores if the document has been parsed

version: WsdlVersion

Stores the WSDL version Note: until the document has been parsed, there is no guarentee that the value stored in this field is correct

types: HashMap<String, WsdlType>

A collection of all the types specified in the WSDL document

Implementations

impl WsdlDocument

pub fn new(path: &Path) -> Result<Self, WsdlError>

Creates a new WsdlDocument structure by reading the contents the file at the provided path

Example

use std::path::Path;
use soap_rs::wsdl::WsdlDocument;

let api_description = WsdlDocument::new(Path::new("test-data/wsdl_sample_1.xml"))
    .expect("Failed to create document");

pub fn parse(&mut self) -> Result<(), WsdlError>

Parses the WSDL document the structure was created with and fills out the structure’s fields to the best of its ability. This is a necessary step in generating the description of a SOAP API and should be called almost always after creating a new structure.

Some special behavior should be noted:

When determining the WSDL document’s version, if there is an error or an unsupported is found, the parser will fall back onto version (WsdlVersion::Wsdl11)[1.1]

Example

use std::path::Path;
use soap_rs::wsdl::WsdlDocument;

let mut api_description = WsdlDocument::new(Path::new("test-data/wsdl_sample_1.xml"))
    .expect("Failed to create document");
if let Err(e) = api_description.parse() {
    eprintln!("Failed to parse WSDL: {}", e);
}